importante é executar a limpeza das tabelas aqui em ambiente de desneolv

+++ prisma/models/city.prisma.ts
@@ -1,11 +1,11 @@
 // prisma/models/city.prisma
 model City {
-  id      Int       @id @default(autoincrement())
-  name    String
-  slug    String    @unique
-  state   State     @relation(fields: [stateId], references: [id])
-  stateId Int
-  stateUf String?
+  id       BigInt    @id @default(autoincrement())
+  name     String
+  slug     String    @unique
+  state    State     @relation(fields: [stateId], references: [id])
+  stateId  BigInt
+  stateUf  String?
   ibgeCode String?   @unique
 
   auctions Auction[]
